+++ khal/cli.py
@@ -258,9 +258,11 @@ def _get_cli():
                   help=('The format of the events.'))
     @click.option('--day-format', '-df',
                   help=('The format of the day line.'))
-    @click.option('--once', '-o', help=('Print event only once'),
-                  is_flag=True)
-    @click.option('--notstarted', help=('Print only events that have not started'),
+    @click.option(
+        '--once', '-o',
+        help=('Print each event only once (even if it is repeated or spans multiple days).'),
+        is_flag=True)
+    @click.option('--notstarted', help=('Print only events that have not started.'),
                   is_flag=True)
     @click.argument('DATERANGE', nargs=-1, required=False)
     @click.pass_context
@@ -297,10 +299,11 @@ def _get_cli():
                   help=('The format of the events.'))
     @click.option('--day-format', '-df',
                   help=('The format of the day line.'))
-    @click.option(
-        '--once', '-o', is_flag=True,
-        help='Print events only once, even if they span multiple days')
-    @click.option('--notstarted', help=('Print only events that have not started'),
+    @click.option('--once', '-o', is_flag=True,
+                  help=('Print each event only once '
+                        '(even if it is repeated or spans multiple days).')
+                  )
+    @click.option('--notstarted', help=('Print only events that have not started.'),
                   is_flag=True)
     @click.argument('DATERANGE', nargs=-1, required=False,
                     metavar='[DATETIME [DATETIME | RANGE]]')
@@ -581,7 +584,7 @@ def _get_cli():
     @click.argument('DATETIME', nargs=-1, required=False, metavar='[DATETIME]')
     @click.pass_context
     def at(ctx, datetime, notstarted, format, day_format):
-        '''Print all events at a specific (date-)time (defaults to now).'''
+        '''Print all events at a specific datetime (defaults to now).'''
         if not datetime:
             datetime = ("now",)
         try:
@@ -589,7 +592,7 @@ def _get_cli():
                 build_collection(ctx.obj['conf'], ctx.obj.get('calendar_selection', None)),
                 agenda_format=format,
                 day_format=day_format,
-                daterange=datetime + ("1m", ),
+                datepoint=list(datetime),
                 once=True,
                 notstarted=notstarted,
                 conf=ctx.obj['conf'],